feat: introduce a multiple appearences signer
As a requirement for signing multiple pages that points to the same
signature, we need to setup a signer that allows for signing the defined
pages.

We overwrite the `CreateNewSignatureFormField` to apply the same
signature to each defined page with the defined position and name.
This allows us to have multiple widgets _without_ signing multiple
times.

Goal is to rely on existing iText functionality to avoid leaking inner
workings between the original `PdfSigner` and the inherited
`MultipleAppearancesPdfSigner`.